+++ api/dashboard/achievement/achievement_serializer.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+from db.achievement import Achievement
+# from db.user import User
+
+class AchievementSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Achievement
+        fields = '__all__'
+
+
