--- api/integrations/qseverse/qseverse_serializer.py
@@ -1,11 +1,0 @@
-from rest_framework import serializers
-from db.achievement import Achievement
-# from db.user import User
-
-class AchievementSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Achievement
-        fields = '__all__'
-
-
